requirements.txt
Django==2.2.3
djangorestframework==3.10.1
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0